// models/request/receivingInstitutionInfoRequest.ts
export interface ReceivingInstitutionInfo {
    receivingInstitutionInfo_id: number;
    university_id: number;
    universityDepartment_id: number;
    academicYear_id: number;
    academicPersonnelName: string;
    academicPersonnelSurname: string;
    academicPersonnelEmail: string;
    phoneNumberE164: string;
    phoneNumberExt: string;
}